requirements.txt
anthropic>=0.40.0
PyYAML>=6.0.2
pandas>=2.2.0
matplotlib>=3.8.0
scipy>=1.12.0


